demo/tech2/requirements.txt
pandas
FinanceDataReader
python-dotenv
openai